pyproject.toml
[project]
name = "pyramid-jwt2"
version = "1.0.0"
description = "JWT authentication security policy for Pyramid 2.0+"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Team Niteo", email = "info@niteo.co" }]
keywords = ["pyramid", "jwt", "authentication", "security"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Framework :: Pyramid",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Internet :: WWW/HTTP",
  "Topic :: Internet :: WWW/HTTP :: WSGI",
  "Typing :: Typed",
]

dependencies = ["pyramid>=2.0", "pyjwt>=2.0", "zope-interface>=5.0"]

[dependency-groups]
dev = [
  "pytest>=8.0",
  "pytest-cov>=4.0",
  "ruff>=0.8",
]

[project.urls]
Homepage = "https://pypi.org/project/pyramid_jwt2"
Repository = "https://github.com/teamniteo/pyramid_jwt2"
Issues = "https://github.com/teamniteo/pyramid_jwt2/issues"
Changelog = "https://github.com/teamniteo/pyramid_jwt2/releases"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pyramid_jwt2"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-ra --strict-markers --strict-config"

[tool.coverage.run]
source = ["pyramid_jwt2"]
branch = true

[tool.coverage.report]
show_missing = true
fail_under = 100
exclude_lines = [
  "pragma: no cover",
  "raise NotImplementedError",
  "if TYPE_CHECKING:",
  "if t.TYPE_CHECKING:",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "ANN401",  # typing.Any disallowed
  "COM812",  # conflicts with formatter
  "D203",    # one-blank-line-before-class
  "D213",    # multi-line-summary-second-line
  "PLC0415", # `import` should be at the top-level of a file
]

[tool.ruff.lint.pydocstyle]
convention = "pep257"

[tool.ruff.lint.per-file-ignores]
"**/tests/**" = [
  "ARG001",  # Unused function argument
  "DTZ001",  # Naive datetime usage
  "N802",    # Function name should be lowercase
  "N803",    # Argument name should be lowercase
  "PLR2004", # Magic value comparisons
  "S101",    # Use of `assert` detected
  "S105",    # Possible hardcoded password
  "S106",    # Possible hardcoded password assigned to argument
  "SLF001",  # Private member accessed
  "TRY002",  # Create your own exception
]
